Add sorting for channel names and labels in dropdown lists
